adds user group support to group dashboard
