+++ webapp/src/main/webapp/src/app/groups/results/group-assessment.service.ts
@@ -5,6 +5,9 @@ import { ResponseUtils } from '../../shared/response-utils';
 import { DataService } from '../../shared/data/data.service';
 import { catchError, map, mergeMap } from 'rxjs/operators';
 import { ReportingServiceRoute } from '../../shared/service-route';
+import { Observable } from 'rxjs/Observable';
+import { MeasuredAssessment } from '../../dashboard/measured-assessment';
+import { MeasuredAssessmentMapper } from '../../dashboard/measured-assessment.mapper';
 
 const ServiceRoute = ReportingServiceRoute;
 
@@ -72,6 +75,7 @@ export class GroupAssessmentService {
     );
   }
 
+  // TODO move to another service
   getTargetsForAssessment(assessmentId: number) {
     return this.dataService.get(`${ServiceRoute}/assessment-targets`, {
       params: {
