--- webapp/src/main/webapp/src/app/groups/results/user-group-assessment.provider.ts
@@ -1,44 +1,0 @@
-import { AssessmentProvider } from '../../assessments/assessment-provider.interface';
-import { Observable } from 'rxjs/Observable';
-import { AssessmentItem } from '../../assessments/model/assessment-item.model';
-import { Exam } from '../../assessments/model/exam.model';
-import { GroupAssessmentService } from './group-assessment.service';
-import { Assessment } from '../../assessments/model/assessment.model';
-import { TargetScoreExam } from '../../assessments/model/target-score-exam.model';
-
-export interface StateProvider {
-  group: { id, schoolId? };
-  schoolYear: number;
-}
-
-export class UserGroupAssessmentProvider implements AssessmentProvider {
-
-  constructor(private service: GroupAssessmentService,
-              private stateProvider: StateProvider) {
-  }
-
-  getAssessmentItems(assessmentId: number, itemTypes?: string[]): Observable<AssessmentItem[]> {
-    const { group, schoolYear } = this.stateProvider;
-    return this.service.getAssessmentItems({userGroupId: group.id, schoolYear, assessmentId, itemTypes});
-  }
-
-  getAvailableAssessments(): Observable<Assessment[]> {
-    const { group, schoolYear } = this.stateProvider;
-    return this.service.getAvailableAssessments({userGroupId: group.id, schoolYear});
-  }
-
-  getExams(assessmentId: number): Observable<Exam[]> {
-    const { group, schoolYear } = this.stateProvider;
-    return this.service.getExams({userGroupId: group.id, schoolYear, assessmentId});
-  }
-
-  getTargetScoreExams(assessmentId: number): Observable<TargetScoreExam[]> {
-    const { group, schoolYear } = this.stateProvider;
-    return this.service.getTargetScoreExams({userGroupId: group.id, schoolYear, assessmentId});
-  }
-
-  getSchoolId(): number {
-    return undefined;
-  }
-
-}
